+++ stable_baselines/common/distributions.py
@@ -1,18 +1,9 @@
-import warnings
-
 import numpy as np
 import tensorflow as tf
 from tensorflow.python.ops import math_ops
 from gym import spaces
 
 from stable_baselines.a2c.utils import linear
-
-
-def warn_out_of_bound():
-    # Warn users that action is not bounded
-    message = "Actions are not clipped and can be out of bound "
-    message += "(see issue https://github.com/hill-a/stable-baselines/issues/36)"
-    warnings.warn(message, stacklevel=2)
 
 
 class ProbabilityDistribution(object):
@@ -401,15 +392,8 @@ class DiagGaussianProbabilityDistribution(ProbabilityDistribution):
         return self.flat
 
     def mode(self):
-        # TODO: squash the ouput instead of clipping
-        # and take the change of distribution into account
-        warn_out_of_bound()
+        # Bounds are taken into account outside this class (during training only)
         return self.mean
-        # low = self.bounds[0]
-        # high = self.bounds[1]
-        #
-        # # clip the output (clip_by_value does not broadcast correctly)
-        # return tf.minimum(tf.maximum(self.mean, low), high)
 
     def neglogp(self, x):
         return 0.5 * tf.reduce_sum(tf.square((x - self.mean) / self.std), axis=-1) \
@@ -425,15 +409,9 @@ class DiagGaussianProbabilityDistribution(ProbabilityDistribution):
         return tf.reduce_sum(self.logstd + .5 * np.log(2.0 * np.pi * np.e), axis=-1)
 
     def sample(self):
-        # TODO: squash the ouput instead of clipping
-        # and take the change of distribution into account
-        warn_out_of_bound()
+        # Bounds are taken into acount outside this class (during training only)
+        # Otherwise, it changes the distribution and breaks PPO2 for instance
         return self.mean + self.std * tf.random_normal(tf.shape(self.mean), dtype=self.mean.dtype)
-        # low = self.bounds[0]
-        # high = self.bounds[1]
-        #
-        # # clip the output (clip_by_value does not broadcast correctly)
-        # return tf.minimum(tf.maximum(self.mean + self.std * tf.random_normal(tf.shape(self.mean)), low), high)
 
     @classmethod
     def fromflat(cls, flat, bounds=(-np.inf, np.inf)):