+++ docs/misc/changelog.rst
@@ -11,7 +11,8 @@ Pre Release 2.0.1.a0 (WIP)
 **logging and bug fixes**
 
 - added patch fix for equal function using `gym.spaces.MultiDiscrete` and `gym.spaces.MultiBinary`
-- temporary removal of action clipping for continuous actions to avoid NaN (see issue #36)
+- removed action clipping for prediction of continuous actions (see issue #36)
+- fixed NaN issue due to clipping the continuous action in the wrong place (issue #36)
 
 
 Release 2.0.0 (2018-09-18)